# docs/api/openapi/member.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Member Service API",
    "description": "회원 관리 서비스 API 문서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway"
    },
    {
      "url": "http://localhost:8083",
      "description": "로컬 개발 서버 (직접 접근)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/member-reports": {
      "post": {
        "tags": [
          "회원 신고"
        ],
        "summary": "회원 신고 생성",
        "description": "회원에 대한 신고를 생성합니다.",
        "operationId": "createReport",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberReportResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/password-resets": {
      "post": {
        "tags": [
          "비밀번호 재설정"
        ],
        "summary": "비밀번호 재설정 메일 발송",
        "description": "입력한 이메일로 비밀번호 재설정 링크를 발송합니다.",
        "operationId": "sendPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePasswordResetSendResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password-resets/confirm": {
      "post": {
        "tags": [
          "비밀번호 재설정"
        ],
        "summary": "비밀번호 재설정 확인",
        "description": "재설정 토큰과 새 비밀번호로 비밀번호를 변경합니다.",
        "operationId": "confirmPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePasswordResetConfirmResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/member-restrictions": {
      "get": {
        "tags": [
          "회원 제재"
        ],
        "summary": "전체 회원 제재 조회",
        "description": "관리자가 전체 회원의 제재 이력을 조회합니다.",
        "operationId": "getAllMemberRestrictions",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMemberRestrictionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "회원 제재"
        ],
        "summary": "회원 제재 생성",
        "description": "회원 제재를 생성합니다.",
        "operationId": "createRestriction",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRestrictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberRestrictionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/members/{memberId}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 조회",
        "description": "회원 프로필",
        "operationId": "getMember",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "description": "회원 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-1111-1111-111111111111"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "404": {
            "description": "회원 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/members/me": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "내 정보 조회",
        "description": "프로필 조회",
        "operationId": "getCurrentMember",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          },
          "404": {
            "description": "회원 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "회원"
        ],
        "summary": "회원 탈퇴",
        "description": "내 계정",
        "operationId": "withdrawCurrentMember",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWithdrawMemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWithdrawMemberResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWithdrawMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "회원"
        ],
        "summary": "내 정보 수정",
        "description": "프로필 수정",
        "operationId": "updateCurrentMember",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/member-restrictions/{restrictionId}/deactivate": {
      "patch": {
        "tags": [
          "회원 제재"
        ],
        "summary": "회원 제재 비활성화",
        "description": "기존 회원 제재를 비활성화합니다.",
        "operationId": "deactivateRestriction",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          },
          {
            "name": "restrictionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMemberRestrictionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/member-reports/me": {
      "get": {
        "tags": [
          "회원 신고"
        ],
        "summary": "내 신고 조회",
        "description": "현재 회원이 작성한 신고를 조회합니다.",
        "operationId": "getMyReports",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMemberReportResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/member-restrictions/members/{memberId}": {
      "get": {
        "tags": [
          "회원 제재"
        ],
        "summary": "회원 제재 조회",
        "description": "특정 회원의 모든 제재를 조회합니다.",
        "operationId": "getMemberRestrictions",
        "parameters": [
          {
            "name": "authenticatedMember",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthenticatedMember"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMemberRestrictionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "AuthenticatedMember": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "SELLER",
              "ADMIN"
            ]
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateMemberReportRequest": {
        "type": "object",
        "properties": {
          "reportedMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "reportType": {
            "type": "string",
            "enum": [
              "ABUSE",
              "FRAUD",
              "SPAM",
              "ETC"
            ]
          }
        },
        "required": [
          "reportType",
          "reportedMemberId"
        ]
      },
      "ApiResponseMemberReportResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/MemberReportResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "MemberReportResponse": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string",
            "format": "uuid"
          },
          "reporterId": {
            "type": "string",
            "format": "uuid"
          },
          "reporterNickname": {
            "type": "string"
          },
          "reportedMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "reportedMemberNickname": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "reportType": {
            "type": "string",
            "enum": [
              "ABUSE",
              "FRAUD",
              "SPAM",
              "ETC"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "reviewComment": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "string",
            "format": "uuid"
          },
          "reviewedByNickname": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PasswordResetSendRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email"
        ]
      },
      "ApiResponsePasswordResetSendResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PasswordResetSendResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "PasswordResetSendResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "PasswordResetConfirmRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          },
          "newPassword": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8
          }
        },
        "required": [
          "newPassword",
          "token"
        ]
      },
      "ApiResponsePasswordResetConfirmResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PasswordResetConfirmResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "PasswordResetConfirmResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateMemberRestrictionRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "restrictionType": {
            "type": "string",
            "enum": [
              "TRADE_BAN",
              "LOGIN_BAN",
              "CHAT_BAN"
            ]
          },
          "durationHours": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "memberId",
          "restrictionType"
        ]
      },
      "ApiResponseMemberRestrictionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/MemberRestrictionResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "MemberRestrictionResponse": {
        "type": "object",
        "properties": {
          "restrictionId": {
            "type": "string",
            "format": "uuid"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "memberNickname": {
            "type": "string"
          },
          "adminId": {
            "type": "string",
            "format": "uuid"
          },
          "adminNickname": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "restrictionType": {
            "type": "string",
            "enum": [
              "TRADE_BAN",
              "LOGIN_BAN",
              "CHAT_BAN"
            ]
          },
          "durationHours": {
            "type": "integer",
            "format": "int32"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateMemberRequest": {
        "type": "object",
        "description": "회원 수정 요청",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "닉네임",
            "example": "updated-user",
            "maxLength": 30,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "description": "전화번호",
            "example": "010-9876-5432"
          },
          "address": {
            "type": "string",
            "description": "주소",
            "example": "부산"
          },
          "profileImageKey": {
            "type": "string",
            "description": "프로필 이미지 키",
            "example": "members/profile/11111111/avatar.png"
          }
        },
        "required": [
          "nickname"
        ]
      },
      "MemberResponse": {
        "type": "object",
        "description": "회원 프로필 응답",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid",
            "description": "회원 ID"
          },
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "member@example.com"
          },
          "nickname": {
            "type": "string",
            "description": "닉네임",
            "example": "goods-user"
          },
          "phone": {
            "type": "string",
            "description": "전화번호",
            "example": "010-1234-5678"
          },
          "address": {
            "type": "string",
            "description": "주소",
            "example": "서울"
          },
          "profileImageUrl": {
            "type": "string",
            "description": "프로필 이미지 URL"
          },
          "role": {
            "type": "string",
            "description": "회원 역할",
            "enum": [
              "USER",
              "SELLER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "description": "회원 상태",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "SUSPENDED",
              "WITHDRAWN",
              "DELETED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "생성 시각"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "수정 시각"
          }
        }
      },
      "ApiResponseMemberResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
...
      "ApiResponseListMemberReportResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberReportResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
...
      "ApiResponseListMemberRestrictionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRestrictionResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "WithdrawMemberRequest": {
        "type": "object",
        "description": "회원 탈퇴 요청",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "현재 비밀번호",
            "example": "password1234",
            "minLength": 1
          }
        },
        "required": [
          "currentPassword"
        ]
      },
      "ApiResponseWithdrawMemberResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/WithdrawMemberResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "WithdrawMemberResponse": {
        "type": "object",
        "description": "회원 탈퇴 응답",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid",
            "description": "회원 ID",
            "example": "11111111-1111-1111-1111-111111111111"
          },
          "status": {
            "type": "string",
            "description": "탈퇴 후 상태",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "SUSPENDED",
              "WITHDRAWN",
              "DELETED"
            ],
            "example": "WITHDRAWN"
          },
          "withdrawnAt": {
            "type": "string",
            "format": "date-time",
            "description": "탈퇴 시각"
          },
          "message": {
            "type": "string",
            "description": "결과 메시지",
            "example": "회원 탈퇴 완료"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}